corrected navbar
