+++ src/Components/Navbar/Navbar.js
@@ -1,16 +1,17 @@
-import React, { Component } from "react";
+import React, { Component } from 'react';
+import { NavLink } from 'react-router-dom';
+import Nja from '../../images/img.jpg';
+import './Navbar.css';
 
-import "./Navbar.css";
 
-class Navbar extends Component {
-  render() {
-    return (
-      <div>
-        <nav class="main-nav navbar-fixed-top">
-          <div class="container-fluid">
-            <div class="navbar-header">
-              <a class="navbar-brand">TherapyMart</a>
-              <button
+class Navbar extends Component{
+    render() {
+        return (
+            <nav className="nav nav navbar">
+            <div className='navdiv container-fluid'>
+    <div className="navbar-header">
+                <NavLink className="navbarbrand navbar-brand" to="/">TherapyMart</NavLink>
+                {/* <button
                 type="button"
                 class="navbar-toggle"
                 data-toggle="collapse"
@@ -20,48 +21,28 @@ class Navbar extends Component {
                 <span class="icon-bar" />
                 <span class="icon-bar" />
                 <span class="icon-bar" />
-              </button>
-            </div>
-            <div class="collapse navbar-collapse">
-              <ul class="nav navbar-nav ">
-                <li class="active">
-                  <a href="">Home</a>
-                </li>
-                <li>
-                  <a href="">Band</a>
-                </li>
-                <li>
-                  <a href="">Therapist</a>
-                </li>
-                <li>
-                  <a href="">Donate</a>
-                </li>
-                <li>
-                  <a href="">Blog</a>
-                </li>
-                <li>
-                  <a href="">About Us</a>
-                </li>
-                <li>
-                  <a href="">Become a Therapist</a>
-                </li>
-              </ul>
-              <div class="dropdown-menu">
-                <button
-                  class="btn btn-default dropdown-toggle"
-                  type="button"
-                  data-toggle="dropdown"
-                >
-                  Dropdown Example
-                  <span class="caret" />
-                </button>
-              </div>
-            </div>
-          </div>
-        </nav>
-      </div>
-    );
-  }
+              </button> */}
+    </div>
+    <ul className="nav navbar-nav">
+      <li className="active"><NavLink to="/">Home</NavLink></li>
+      <li ><NavLink to="/therapists">Therapists</NavLink></li>
+      <li ><NavLink to="/donate">Donate</NavLink></li>
+      <li ><NavLink to="/blog">Blog</NavLink></li>
+      <li ><NavLink to="/about_us">About Us</NavLink></li>
+      <li ><NavLink to="/new_therepist">Become a Therapist</NavLink></li>
+                </ul>
+              <div className="navright nav-right">
+                    <ul className="nav navbar-nav navbar-right">
+                  <li className="nja">
+                    <img className="nja_img" src={Nja} alt='nigeria_icon' />  &#8358; <span className="caret"></span>
+                    </li>
+                  <li className="acc"><span className="fa fa-user-circle fa-3x fa3x"></span> Account<span className="caret"></span></li>
+                  </ul>
+                
+                    </div>
+  </div>
+</nav>
+        )
+    }
 }
-
 export default Navbar;
